requirements.txt
beautifulsoup4==4.11.1
jsonlines==3.1.0
numpy==1.24.1
pandas==1.5.2
pandas-access==0.0.1
PyYAML==6.0
requests==2.28.1
requests-cache==0.9.7
requests-futures==1.0.0
scipy==1.10.0
python-dotenv==0.21.0
luigi==3.3.0
openpyxl==3.1.2
html5lib==1.1
tqdm==4.65.0
pyarrow==12.0.1